# docs/openapi-core-functions.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SightIndex 核心接口：上传、检索、人脸库",
    "version": "2026-09-09-semantic-search-v1",
    "description": "从北京已部署OpenAPI裁剪，配套api-core-functions.zh-CN.md。核对时间2026-09-09 09:52+08:00。保留原请求/成功响应合同；中文摘要、Basic鉴权和部分错误说明为文档补充，不是新增服务能力。新上传数据自动Qwen索引当前关闭；属性生成模型未启用；人员编辑/删除、人脸删除尚无接口。"
  },
  "servers": [
    {
      "url": "http://192.168.1.40:18030",
      "description": "北京内网示例；导入后替换为调用方实际可达地址，不代表公网可达"
    }
  ],
...
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
...
    "/api/images/upload": {
      "post": {
        "tags": [
          "上传"
        ],
        "summary": "上传图片",
        "operationId": "upload_image_api_images_upload_post",
        "parameters": [
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "upload",
              "title": "Source Type"
            }
          },
          {
            "name": "camera_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Camera Id"
            }
          },
          {
            "name": "location_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          },
          {
            "name": "captured_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Captured At"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_api_images_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "description": "模型/索引未就绪或不可用、处理队列压力等。视频可能有已提交进度；不能把此状态当作正常空匹配。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "object",
                          "additionalProperties": true
                        },
                        {
                          "type": "array",
                          "items": {}
                        }
                      ]
                    }
                  },
                  "required": [
                    "detail"
                  ]
                }
              }
            }
          }
        },
        "description": "file 是 multipart 文件，其余元数据为 query 参数。保存原图不等于生成人物裁剪或完成索引；非幂等。"
      }
    },
    "/api/images": {
      "get": {
        "tags": [
          "上传"
        ],
        "summary": "图片列表",
        "operationId": "list_images_api_images_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "has_crops",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Has Crops"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageRead"
                  },
                  "title": "Response List Images Api Images Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "返回裸数组；limit 1～200，支持 offset 和 has_crops。"
      }
    },
    "/api/images/{image_id}": {
      "get": {
        "tags": [
          "上传"
        ],
        "summary": "图片详情",
        "operationId": "get_image_api_images__image_id__get",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "路径中的业务对象不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "object",
                          "additionalProperties": true
                        },
                        {
                          "type": "array",
                          "items": {}
                        }
                      ]
                    }
                  },
                  "required": [
                    "detail"
                  ]
                }
              }
            }
          }
        },
        "description": "image_url 等为受认证保护的相对资源路径。"
      }
    },
    "/api/media/counts": {
      "get": {
        "tags": [
          "上传"
        ],
        "summary": "媒体计数",
        "operationId": "get_media_counts_api_media_counts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaCounts"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "image_with_crops_count 是有裁剪的图片数，不是全部上传图片数。"
      }
    },
    "/api/person-crops": {
      "get": {
        "tags": [
          "上传"
        ],
        "summary": "人物裁剪列表",
        "operationId": "list_person_crops_api_person_crops_get",
        "parameters": [
          {
            "name": "image_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Image Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonCropRead"
                  },
                  "title": "Response List Person Crops Api Person Crops Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "返回裸数组，可按 image_id 查询。"
      }
    },
    "/api/person-crops/{crop_id}": {
      "get": {
        "tags": [
          "上传"
        ],
        "summary": "人物裁剪详情",
        "description": "人员关联可能为空，裁剪框基于原图像素坐标。",
        "operationId": "get_person_crop_api_person_crops__crop_id__get",
        "parameters": [
          {
            "name": "crop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Crop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonCropRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "路径中的业务对象不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "object",
                          "additionalProperties": true
                        },
                        {
                          "type": "array",
                          "items": {}
                        }
                      ]
                    }
                  },
                  "required": [
                    "detail"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/images/{image_id}/process": {
      "post": {
        "tags": [
          "上传"
        ],
        "summary": "处理图片并生成人物裁剪",
        "operationId": "process_image_api_images__image_id__process_post",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Image Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonCropRead"
                  },
                  "title": "Response Process Image Api Images  Image Id  Process Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Basic认证失败；当前响应可为纯文本，并带WWW-Authenticate。",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "路径中的业务对象不存在。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "object",
                          "additionalProperties": true
                        },
                        {
                          "type": "array",
                          "items": {}
                        }
                      ]
                    }
                  },
                  "required": [
                    "detail"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "模型/索引未就绪或不可用、处理队列压力等。视频可能有已提交进度；不能把此状态当作正常空匹配。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "object",
                          "additionalProperties": true
                        },
                        {
                          "type": "array",
                          "items": {}
                        }
                      ]
                    }
                  },
                  "required": [
                    "detail"
                  ]
                }
              }
            }
          }
        },
        "description": "返回裸数组。当前北京版本重复调用可能新增重复裁剪，无通用幂等保证。200 不保证可选属性和索引步骤全部完成。"
      }
    },
...
  },
  "components": {
    "schemas": {
      "MediaCounts": {
        "properties": {
          "image_with_crops_count": {
            "type": "integer",
            "title": "Image With Crops Count"
          },
          "person_crop_count": {
            "type": "integer",
            "title": "Person Crop Count"
          }
        },
        "type": "object",
        "required": [
          "image_with_crops_count",
          "person_crop_count"
        ],
        "title": "MediaCounts",
        "description": "Untruncated totals for the monitor/search overview."
      },
      "Body_upload_image_api_images_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_image_api_images_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PersonCropRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "image_id": {
            "type": "string",
            "format": "uuid",
            "title": "Image Id"
          },
          "crop_url": {
            "type": "string",
            "title": "Crop Url"
          },
          "bbox": {
            "additionalProperties": true,
            "type": "object",
            "title": "Bbox"
          },
          "attributes": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "camera_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Id"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "image_id",
          "crop_url",
          "bbox",
          "person_id",
          "camera_id",
          "location_id",
          "captured_at",
          "created_at"
        ],
        "title": "PersonCropRead"
      },
...
      "ImageRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "camera_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Id"
          },
          "location_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Name"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "image_url",
          "thumbnail_url",
          "source_type",
          "camera_id",
          "location_id",
          "captured_at",
          "created_at"
        ],
        "title": "ImageRead"
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "实际部署由中间件执行，账号密码单独配置，本文不包含凭据。"
      }
    }
  },
...
}